pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf_pilot"
version = "0.1.0"
description = "The only PDF converter that automatically selects the best engine for each document"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "pdf_pilot contributors"}
]
keywords = [
    "pdf", "pdf-converter", "pdf-parser", "pdf-to-markdown",
    "document-parsing", "document-analysis", "ocr",
    "layout-analysis", "table-extraction", "rag", "llm",
    "markdown", "word", "docx",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business :: Office Suites",
]
requires-python = ">=3.10"
dependencies = [
    "pymupdf4llm>=1.0",
    "python-docx>=1.0",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
mineru = ["magic-pdf>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-html>=4.0",
    "ruff>=0.4.0",
]

[project.scripts]
pdf_pilot = "pdf_pilot.cli:main"

[project.urls]
Homepage = "https://github.com/Electricitysheep/pdf_pilot"
Repository = "https://github.com/Electricitysheep/pdf_pilot"
Issues = "https://github.com/Electricitysheep/pdf_pilot/issues"
Changelog = "https://github.com/Electricitysheep/pdf_pilot/blob/main/CHANGELOG.md"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
